+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 * synth support for variable frame rates
 * Various documentation updates
 * Cross-synthesis capability added with pyatsyn.analysis.merge
+* synthesis now allows an overriding sampling rate to be specified
 
 
 [v1.0.0](https://github.com/boxofbox/pyatsyn/releases/tag/v1.0.0) Initial Release (2023-01-22)